refactor: move roles generator into lib
