+++ lib/roles.js
